+++ js/changePictureScale.js
@@ -1,0 +1,1 @@
+//Функция изменения масштаба изображения.
