+++ js/clearForm.js
@@ -1,0 +1,1 @@
+//функция очистки формы при успешной отправки запроса на сервер. Появление сообщения об успешной/неуспешной отправках.
