+++ js/displayComments.js
@@ -1,0 +1,1 @@
+//функция, выводящая комментарии к изображению.
