+++ js/editPictureForm.js
@@ -1,0 +1,1 @@
+//Функция открытия/закрытия модального окна редактирования фотографии
