+++ js/errorMessage.js
@@ -1,0 +1,1 @@
+//функция выводящая сообщение об ошибке отправки изображения с сервера.
