+++ js/idNum.js
@@ -1,0 +1,5 @@
+//Массив из 25 чисел
+const idNum = new Array(25).fill(null);
+for (let int = 0; int < idNum.length - 1; int++) {
+  idNum.push(int);
+}
