+++ js/imposeEffectOnPicture.js
@@ -1,0 +1,1 @@
+//Функция наложения эффектов на фото
