+++ js/interactionPicture.js
@@ -1,0 +1,1 @@
+//Функция отвечающая за открытие и закрытие фото.
