+++ js/openMiniPhoto.js
@@ -1,0 +1,1 @@
+//Открытие миниатюры.
