+++ js/slider.js
@@ -1,0 +1,1 @@
+//функция взаимодействия со слайдером
