+++ js/uploadNewPhoto.js
@@ -1,0 +1,1 @@
+//Функция загрузки нового изображения
